V2.2

+++ PLink/app/Models/SmartBin.php
@@ -21,7 +21,7 @@ class SmartBin extends Model
     ];
     protected $casts = [
         'current_fill_percentage' => 'integer',
-        'current_distance_cm' => 'integer',
+        'current_distance_cm' => 'float',
         'full_threshold_cm' => 'integer',
         'empty_threshold_cm' => 'integer',
         'last_maintenance_at' => 'datetime',
